Include <stdint.h> where needed
The {u,}int{8,16,32}_t types are defined in <stdint.h>, so it should be
included when such types are used.

Not including <stdint.h> might work by accident with some C libraries
due to it being included by other headers, but it for example causes
build failures with the musl C library.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>